+++ click_and_drop_api/examples/__init__.py
@@ -7,6 +7,7 @@ Each module can be executed directly::
     python -m click_and_drop_api.examples.shipping_countries
     python -m click_and_drop_api.examples.create_order
     python -m click_and_drop_api.examples.view_specific_orders
+    python -m click_and_drop_api.examples.view_order
     python -m click_and_drop_api.examples.package_size_and_cost
     python -m click_and_drop_api.examples.generate_label
 """